Streaming and also move to new file for clarity
